requirements.txt
allure-behave==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
axe-selenium-python==2.1.6
azure-functions==1.15.0
behave==1.2.6
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
cryptography==41.0.1
cssselect2==0.7.0
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
Faker==18.11.2
h11==0.14.0
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
MouseInfo==0.1.3
mss==9.0.1
mysql-connector-python==8.0.33
numpy==1.25.0
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parse==1.19.1
parse-type==0.6.1
Pillow==10.0.0
pluggy==1.2.0
protobuf==3.20.3
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyodbc==4.0.39
PyPDF2==3.0.1
pyperclip==1.8.2
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
PySocks==1.7.1
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytweening==1.0.7
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
svglib==1.5.1
tinycss2==1.2.1
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.3
webdriver-manager==3.8.6
webencodings==0.5.1
wsproto==1.2.0
xmltodict==0.13.0